LPS-123455 Convert the deleteKaleoProcess action method to a mvcActionCommand and apply to usages
